test(batch-auth): dissolve extended test file into proper locations
- New store tests (removeSlot statuses, excelError field, retryFailed,
  isBusy, addBlockedPort, resetAuthStats, checkBatchCompletion gate,
  web-mode no-ops, dismissBanner, currentStats) → batch-flash-auth.test.ts
- TC-061 downloadAuthFirmware web-mode guard → auth-firmware.test.ts
- port-filter and filterByChip tests were already covered by their
  co-located test files; duplicate sections dropped
- Delete batch-flash-auth-extended.test.ts (non-standard naming)
